% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{Number of trials}

\item{k}{Number of successes}
}
\value{
Number of possible ways you can can n trials with k successes
}
\description{
Calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(n = 5, k = 0)
}
