% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Three column data frame with two classes c("bincum", "data.frame") specifying cumulative probabilities
}
\description{
Calculates the cumulative probability for the binomial distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
bin_cumulative(trials = 10, prob = 0.1)
}
