% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Two column data frame with two classes c("bindis", "data.frame") specifying probability
}
\description{
Calculates the probability dist of varying number of success with specified number of trials and success prob
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
bin_distribution(trials = 10, prob = 0.1)
}
