% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Kurtosis of the binomial distribution with a specified number of trials and success probability
}
\description{
Calculates the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
bin_kurtosis(100, 0.1)
}
