% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_mode}
\alias{bin_mode}
\title{Binomial Mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Mode of the binomial distribution with a specified number of trials and success probability
}
\description{
Calculates the mode of a binomial distribution
}
\examples{
bin_mode(10, 0.3)
bin_mode(100, 0.1)
}
