% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Binomial random variable object of class "binvar"
}
\description{
Creates a binomial variable with specified number of trials and success probability
}
\examples{
bin_var <- bin_variable(trials = 5, prob = 0.5)
bin_var
}
