% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_func.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial Variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{Number of trials}

\item{prob}{Probability of success}
}
\value{
Variance of the binomial distribution with a specified number of trials and success probability
}
\description{
Calculates the variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
bin_variance(100, 0.1)
}
